optimize testfiles
