--- simulator/spec/circuits/nandGate.json
@@ -1,1 +1,0 @@
-{"name":"not gate","timePeriod":500,"clockEnabled":true,"projectId":"yA5ma9nWFLcsOq9kQYeS","focussedCircuit":78657923676,"orderedTabs":["78657923676"],"scopes":[{"layout":{"width":100,"height":60,"title_x":50,"title_y":13,"titleEnabled":true},"verilogMetadata":{"isVerilogCircuit":false,"isMainCircuit":false,"code":"// Write Some Verilog Code Here!","subCircuitScopeIds":[]},"allNodes":[{"x":10,"y":0,"type":0,"bitWidth":1,"label":"","connections":[5]},{"x":10,"y":0,"type":1,"bitWidth":1,"label":"","connections":[3]},{"x":10,"y":0,"type":1,"bitWidth":1,"label":"","connections":[4]},{"x":-10,"y":-10,"type":0,"bitWidth":1,"label":"","connections":[1]},{"x":-10,"y":10,"type":0,"bitWidth":1,"label":"","connections":[2]},{"x":30,"y":0,"type":1,"bitWidth":1,"label":"","connections":[0]}],"testbenchData":{"currentCase":0,"currentGroup":0,"testData":{"type":"comb","title":"Untitled","groups":[{"label":"Group 1","inputs":[{"label":"inp1","bitWidth":1,"values":["1","0"]}],"outputs":[{"label":"out1","bitWidth":1,"values":["0","1"],"results":["0","1"]}],"n":2}]}},"id":78657923676,"name":"Main","Input":[{"x":-170,"y":-110,"objectType":"Input","label":"inp1","direction":"RIGHT","labelDirection":"LEFT","propagationDelay":0,"customData":{"nodes":{"output1":1},"values":{"state":1},"constructorParamaters":["RIGHT",1,{"x":0,"y":20,"id":"XYbrP4LqVcwEu75NulQB"}]}},{"x":-170,"y":-90,"objectType":"Input","label":"inp2","direction":"RIGHT","labelDirection":"LEFT","propagationDelay":0,"customData":{"nodes":{"output1":2},"values":{"state":1},"constructorParamaters":["RIGHT",1,{"x":0,"y":40,"id":"6LoqU0kqpxaEP1JTTcAA"}]}}],"Output":[{"x":80,"y":-100,"objectType":"Output","label":"out1","direction":"LEFT","labelDirection":"RIGHT","propagationDelay":0,"customData":{"nodes":{"inp1":0},"constructorParamaters":["LEFT",1,{"x":100,"y":20,"id":"v8KvEGsGwuzwcjxr39Ug"}]}}],"NandGate":[{"x":-40,"y":-100,"objectType":"NandGate","label":"nandgate","direction":"RIGHT","labelDirection":"LEFT","propagationDelay":10,"customData":{"constructorParamaters":["RIGHT",2,1],"nodes":{"inp":[3,4],"output1":5}}}],"restrictedCircuitElementsUsed":[],"nodes":[]}]}
